Create ubuntu-basic
The following script is an example of how to execute commands in Ubuntu Linux through xterm. The script first initiates the run application window through ubuntu and later on it triggers the command xterm. After the execution of the terminal it issues the commands "pwd", "id" and "cat /etc/passwd".